requirements.txt
Flask
flask-socketio
opencv-python-headless
ultralytics
Pillow
